requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
yt-dlp>=2023.3.4
requests>=2.27.1
TTS>=0.8.0
ffmpeg-python>=0.2.0
asyncio>=3.4.3
uuid>=1.30 